[hyeok][refactor] signup rendering refactor
